Refactor developer database table
We don't need a separate integer primary key for anything yet, so we can
just use the username as the primary key. This also makes
de/serialization easier to work with.

We also don't need the database to ensure public keys are unique. Public
keys' uniqueness should be validated before they're ever uploaded to the
app repository, and if the repodata is inconsistent, then the app can't
really do anything about that anyway.