Set baseline constraints for background update job
The network must be connected for app updates, the device should be idle
to prevent stopping apps in active use, and the device storage must be
high enough to download apps.

Other constraints (e.g. not updating on metered networks) may be added
in the future, but they will likely be user-configurable. These three
are applicable to all users.