Define backup rules
android:allowBackup is deprecated as of Android 12 and it only applies
to cloud storage, so we should manually define backup rules for cloud
backup and D2D transfers separately.

For now we'll play it safe and disable cloud backup if it has no
encryption capabilities. Once the app stabilizes, we can evaluate
whether we want to loosen this restriction.