Add back missing app label
In ccb59c9, I meant to remove the app label from the activity
declaration in the manifest since it's not necessary when the label is
specified for the application. I accidentally removed the label for the
application instead of the activity, causing the app to display its name
as app.accrescent.client.Accrescent.