Use padding param in Scaffold content lambda
The lint warning explains why this is important pretty well, so I've
copied it below:

The content lambda in Scaffold has a padding parameter which will
include any inner padding for the content due to app bars. If this
parameter is ignored, then content may be obscured by the app bars
resulting in visual issues or elements that can't be interacted with.

This lint was introduced in compose-ui 1.2.0-alpha07.
https://developer.android.com/jetpack/androidx/releases/compose-ui#1.2.0-alpha07.