Verify hashes of signing certs instead of public keys
The Android platform recognizes a certificate identity by the hash of
its contents. Accrescent is currently treating the hash of certificate's
public key as its identity, which is a more relaxed interpretation of
certificate identities. We have no reason to deviate from Android in
this regard and Android's interpretation is more specific anyway, so we
should play it safe and identify app signers by the hash of whole
certificates.

This is a breaking change to the signed repodata format and will require
a manual update.

Closes accrescent/meta#6