Add "Open" button for when app is already installed
When an app is installed, the "Install" button in the app list will
change to "Open" and pressing it will open the respective app.

AppInstallStatuses was made into a singleton injectable because it will
be needed later in other screens such as the app detail screen and
update list screen.