Support repodata signing key rotation
Key can be rotated by:

1. Updating REPODATA_PUBKEY
2. Bumping PUBKEY_VERSION
3. Creating new signed metadata at repodata.$PUBKEY_VERSION.json
4. Creating new metadata sigs at repodata.$PUBKEY_VERSION.json.sig

This change is sort of backwards incompatible since it assumes a
different remote path for the repodata, but since the server can
continue to serve repodata at the old path, older clients will continue
to work until it is removed.

We'll leave the old repodata up for a few weeks to allow more people to
automatically upgrade.

Closes #46