Improve security of repodata downgrade protection
Previously MIN_TIMESTAMP was only used as a fallback when a timestamp
was not previously saved. This behavior means that MIN_TIMESTAMP is not
respected when a saved timestamp is present. In practice, this is not
usually an issue since a first-time install of Accrescent won't save a
timestamp less than MIN_TIMESTAMP. However, if repodata updates are
blocked for one reason or another and the user sideloads an Accrescent
update with a newer MIN_TIMESTAMP, MIN_TIMESTAMP could be greater than
the saved timestamp but the saved timestamp will be incorrectly treated
as the source of truth.

This commit ensures than MIN_TIMESTAMP is truly treated as a minimum
value rather than just a fallback.