Target SDK 33
This includes some basic logic for requesting the notification
permission on MainActivity creation if it hasn't been granted already.
Admittedly this solution isn't ideal if the user wants to deny the
notification permission, but I've decided to delegate implementing
another way to handle STATUS_PENDING_USER_ACTION to a later commit so we
can target SDK 33 now.

Note there is no permission rationale shown for the same reason, but
this isn't a regression so it's less of a concern (although we certainly
want to add permission rationale eventually).