Only install language split APKs opportunistically
Requiring a language split APK to exist when installing an app is
unnecessary since default string resources are included in the base APK.
In fact, because default string resources are included in the base APK,
removing unused alternative resources as per
https://developer.android.com/studio/build/shrink-code#unused-alt-resources
will block generation of a split APK for the default language.

The end result of this behavior is that Accrescent will fail to install
an app whose default language matches the preferred device language if
said app removes unused alternative resources (which is rather ironic).

As a temporary workaround until clients are updated to include this
patch, the server can offer a valid language split for a different
language when the nonexistent default language split is requested.

Note that the comment about how the Play Store handles language changes
on-demand was removed. This is because the nearby code isn't relevant to
that functionality and the comment needed to be updated anyway to
properly reflect how language splits are handled after this commit.