Add option to prompt user for installs when privileged
When Accrescent is privileged, the OS doesn't prompt the user when
installing an app. This may be undesirable since the user could
accidentally press "Install" on unwanted apps. This commit allows the
user to force OS prompts for installations when Accrescent is
privileged. Note this doesn't affect app updates i.e. pressing an
"Update" button while the option is enabled will not prompt the user.

The option is disabled by default.