Use custom confirmation dialog when privileged
Currently when Accrescent is privileged and confirmation dialogs are
enabled, it passes USER_ACTION_REQUIRED to the package installer session
parameters. However, in testing this approach encountered an issue:
REQUEST_INSTALL_PACKAGES must be granted even if INSTALL_PACKAGES is
already granted. This is a problem for a few reasons:

- The user has to grant Accrescent REQUEST_INSTALL_PACKAGES if a decent
  UX is to be expected
- The user can deny REQUEST_INSTALL_PACKAGES at any time, degrading UX
  for no benefit (since INSTALL_PACKAGES is more powerful anyway)
- This behavior implies that apps installed this way may be considered
  to be from a third-party source, in which case UserManager
  restrictions blocking a user from using third-party app sources would
  incorrectly block apps from Accrescent

The first issue can be solved by having the OS grant Accrescent
REQUEST_INSTALL_PACKAGES by default. The second may even be possible to
prevent in the OS, although I don't know if it's feasible for certain.
But the third, if true, is a major UX downgrade and requires users
blocking third-party app sources to not use the dialog toggle.
Additionally, the OS confirmation dialog prompted by
USER_ACTION_REQUIRED isn't stylistically consistent with Accrescent's
custom uninstall dialog. Even if the formermost two issues are solvable
via OS configuration, this is extra work on behalf of the OS maintainer
and is another hurdle for adoption.

The solution to all of this is quite simple. Instead of passing
USER_ACTION_REQUIRED to the package installer session when dialogs are
enabled, always pass USER_ACTION_NOT_REQUIRED as before and implement a
custom confirmation dialog for installations based on the logic for the
uninstallation confirmation dialog.

Prior discussion:

https://github.com/accrescent/accrescent/pull/74#issuecomment-1319467321
https://github.com/accrescent/accrescent/pull/74#issuecomment-1320633209