Prevent cache bloating
Currently Accrescent downloads APKs to its internal cache directory,
which is all well and good. However, it doesn't attempt to delete them,
meaning that over time, Accrescent begins to take up an exorbitant of
internal storage. We could attempt to delete these APKs when we're done
using them, but the cache bloating problem would still occur if e.g.
Accrecent is stopped or the OS crashes.

Instead, create a new TemporaryFile class which represents storage
backed by a file descriptor but not any particular file (the code for
which is almost entirely derived from GrapheneOS's Apps app). APKs are
now downloaded to this descriptor to avoid any possibility for cache
bloating.

Closes #143