Fix indefinitely loading install status when opening deep link
When opening a deep link in Accrescent when it hasn't previously been
opened to an AppList screen, the internal state of package install
statuses is null for all packages since it is only initialized on the
AppListViewModel initialization. This means that when opening a deep
link directly to the app details screen, the package install status is
never resolved, resulting in an infinitely loading install button.

This logic should ultimately probably be moved outside of the
ViewModels, but for now, fetch the new package install status in the
AppDetailsViewModel as well.