Properly respect network settings in updater work
When the application is started, the network type used for initially
enqueueing periodic update work is incorrectly set to CONNECTED instead
of the preference set by the user. This means that the network type may
be incorrect from the launch of the application until the user sets the
network type again, and it will only be correct until Accrescent is
killed by the OS.

Instead, use the user's preferred network type when initially enqueueing
updater work.