Remove icon hash from repository index
As per https://github.com/accrescent/meta/issues/31, we plan to no
longer include the icon hash for store listings in the signed repository
index, so we should remove it from existing app databases and start
ignoring it for future updates.

Existing clients require this field to exist, so we'll still include it
server-side until most clients are expected to have updated to prevent
breakage.