Disable TLS certificate pinning
We may not have the ability to use Let's Encrypt certificates or our own
backup certificates on the next iteration of Accrescent's
infrastructure. Thus, we have to disable TLS certificate pinning at
least temporarily so that clients can continue to update.

Such a server change will break existing clients which haven't updated,
so we'll wait some time before rolling it out to allow clients to
update.