Change repository base URL
This change is dual-purpose. First, it allows us to reserve
store.accrescent.app for a future web frontend to the store. Second, it
allows us to see from server logs how many clients have updated to
disable TLS certificate pinning as in 51b0f72, helping us determine when
to migrate the repository infrastructure.