Don't automatically updates apps installed via APK
5cac806: "Don't automatically update apps installed by another store"
retained Accrescent's ability to attempt to automatically update apps
installed via the system PackageInstaller, which is the case when
installing an app from an APK file. However, this is not always
desirable, e.g., when an app variant distributed via APK differs from
the variant published in Accrescent.

This change causes Accrescent to skip update attempts for apps installed
via APK except for itself. This exception is necessary to allow
Accrescent to take over updates for itself after being installed from
the website.