Prevent accidentally changing installers of record
Users may unknowingly change the installer of record for an app they
installed from outside of Accrescent by updating it from within
Accrescent, possibly changing the variant they have installed and
disrupting updates from the expected app source.

We should provide a way for users to explicitly choose to change the
installer of record to Accrescent when they want to, but for now, we
should prevent them from doing so accidentally.