Migrate to the directory service
The directory service is now the source of truth for app listings,
available updates, and download information. Now the only use of the
legacy repository model is verifying signed app metadata such as signing
certificates and minimum version codes.

This change duplicates screen UI code quite heavily, which is obviously
not ideal. A more extensive overhaul to the UI code is needed to resolve
this issue cleanly, so it was decided to accept the temporary code
duplication to get the initial directory service migration off the
ground.

Accrescent's package installer no longer respects the app's requested
install location due to additional complexity and the fact that package
installers are allowed to do this according to the platform's API
documentation. We may revert this behavior in the future, but as it is
not critical to keep, this is not a priority.