Remove unnecessary short description null fallback
`listing.shortDescription` will never be null because of the behavior of
the protobuf API, so the present null fallback will never occur.
Additionally, all apps in Accrescent have a short description, and all
future apps are required to have one, so an app's short description will
never be empty either, completely eliminating the need for an emptiness
check.