Fix long URLs breaking app installation
As of 47c6f1b, the APK name passed to `PackageInstaller.openWrite()` is
the hexadecimal representation of the APK's URL. Although the
documentation for that method claims the APK name parameter can be
arbitrary as long as it's unique, it actually has to be a valid ext4
file name. This behavior means that if the hex representation of an APK
url exceeds 255 characters, the APK cannot be installed.

I've filed an upstream AOSP issue at
https://issuetracker.google.com/issues/447174551 since this is at the
very least a documentation bug, but to work around it, we can simply
hash the APK url with SHA-256 to obtain a constant-length APK name which
is always a valid ext4 file name.