Show only user-initiated install workers in UI
Because update workers enqueued by the auto update worker are not
necessarily run immediately, they can appear as pending in their
corresponding app details screens for a very long time. Additionally, if
an update worker enqueued by the auto update worker fails, that error is
persisted until WorkManager reaps it, possibly displaying it to users at
confusing times.

To solve both of these issues, tag user-initiated install workers with a
USER_INITIATED tag and display only these in the UI. That way both auto
update pending statuses and errors are not displayed in the UI.

Fixes #751