Opt out of `GENTLE_UPDATE` for self updates
GENTLE_UPDATE is very passive, by default (at the time of writing) not
updating apps "interacting with the user", including apps that are
"sending or receiving network data". Because self updates are the most
critical path for Accrescent, opt out of GENTLE_UPDATE for self updates
for now so that we don't risk leaving Accrescent users without updates.