Don't use gentle updates for manual updates
Using gentle updates for manual updates results in confusing behavior
wherein manually updating an app may download it completely, but then
appear to fail to install because progress in longer made in the UI.

Disable gentle updates for manual updates to match user-expected
behavior.

Gentle updates are still used for auto updates since those should be as
invisible as possible to the user.

Fixes #760